Split out hasDomainCredit
